fix(app-website-builder): size sidebar scroll by measuring its own top offset
Fixed/token/reserved-space heights all mis-sized the sidebar scroll box: the
shared uiReservedSpace also counts the canvas's address bar and breadcrumbs, which
sit beside the sidebar, not above it. Measure the scroll container's own distance
from the viewport top instead and cap max-height to it — self-correcting regardless
of the chrome above, hugging short content (no gap) and scrolling when tall.
